+++ src/components/dao/widget-roles.vue
@@ -6,38 +6,26 @@ const STATES = Object.freeze({
   WAITING: 'WAITING',
   CREATING_ROLE: 'CREATING_ROLE',
   SAVING_ROLE: 'SAVING_ROLE',
-  CREATING_LEVEL: 'CREATING_LEVEL',
-  SAVING_LEVEL: 'SAVING_LEVEL'
+  CREATING_TIER: 'CREATING_TIER',
+  SAVING_TIER: 'SAVING_TIER'
 })
 
 const TABS = Object.freeze({
   TYPE: 'TYPE',
-  LEVEL: 'LEVEL'
+  TIER: 'TIER'
 })
 
-const LEVEL = {
-  name: '',
-  annualAmount: '',
-  minDeferred: 10
-}
-
+//
 const ROLE = {
   name: '',
   description: ''
 }
 
-const LEVELS_QUERY = `
-  querySalaryband(
-    filter: {
-      details_dao_i: { eq: $daoId },
-    }
-  ) {
-    id: docId
-    name: details_name_s
-    annualAmount: details_annualUsdSalary_a
-    minDeferred: details_minDeferredX100_i
-  }
-`
+const TIER = {
+  name: '',
+  annualAmount: '',
+  minDeferred: 10
+}
 
 const ROLES_QUERY = `
   queryRole(
@@ -53,6 +41,23 @@ const ROLES_QUERY = `
     assignmentAggregate(filter: { 
       details_state_s: { regexp: "/approved/" } 
     }) {
+      count
+    }
+  }
+`
+
+const TIERS_QUERY = `
+  querySalaryband(
+    filter: {
+      details_dao_i: { eq: $daoId },
+    }
+  ) {
+    id: docId
+    name: details_name_s
+    annualAmount: details_annualUsdSalary_a
+    minDeferred: details_minDeferredX100_i
+
+    assignmentAggregate {
       count
     }
   }
@@ -77,13 +82,13 @@ export default {
   },
 
   apollo: {
-    levels: {
-      query: gql`query LEVELS($daoId: Int64!) { ${LEVELS_QUERY} }`,
-      update: data => data.querySalaryband,
+    roles: {
+      query: gql`query ROLES($daoId: Int64!) { ${ROLES_QUERY} }`,
+      update: data => data.queryRole,
       skip () { return !this.selectedDao?.docId },
       variables () { return { daoId: this.selectedDao.docId } },
       subscribeToMore: {
-        document: gql`subscription LEVELS($daoId: Int64!) { ${LEVELS_QUERY} }`,
+        document: gql`subscription ROLES($daoId: Int64!) { ${ROLES_QUERY} }`,
         skip () { return !this.selectedDao?.docId },
         variables () { return { daoId: this.selectedDao.docId } },
         updateQuery: (previousResult, { subscriptionData }) => {
@@ -99,13 +104,13 @@ export default {
       }
     },
 
-    roles: {
-      query: gql`query ROLES($daoId: Int64!) { ${ROLES_QUERY} }`,
-      update: data => data.queryRole,
+    tiers: {
+      query: gql`query TIERS($daoId: Int64!) { ${TIERS_QUERY} }`,
+      update: data => data.querySalaryband,
       skip () { return !this.selectedDao?.docId },
       variables () { return { daoId: this.selectedDao.docId } },
       subscribeToMore: {
-        document: gql`subscription ROLES($daoId: Int64!) { ${ROLES_QUERY} }`,
+        document: gql`subscription TIERS($daoId: Int64!) { ${TIERS_QUERY} }`,
         skip () { return !this.selectedDao?.docId },
         variables () { return { daoId: this.selectedDao.docId } },
         updateQuery: (previousResult, { subscriptionData }) => {
@@ -130,8 +135,8 @@ export default {
       state: STATES.WAITING,
       tab: TABS.TYPE,
 
-      level: { ...LEVEL },
-      role: { ...ROLE }
+      role: { ...ROLE },
+      tier: { ...TIER }
     }
   },
 
@@ -140,8 +145,8 @@ export default {
 
     async _createLevel () {
       try {
-        await this.createLevel({ data: [{ ...this.level }] })
-        this.level = { ...LEVEL }
+        await this.createLevel({ data: [{ ...this.tier }] })
+        this.tier = { ...TIER }
       } catch (e) {
         const message = e.message || e.cause.message
         this.showNotification({ message, color: 'red' })
@@ -150,9 +155,9 @@ export default {
       this.state = STATES.WAITING
     },
 
-    async _deleteLevel (level) {
+    async _deleteLevel (tier) {
       try {
-        await this.deleteLevel({ data: [{ ...level }] })
+        await this.deleteLevel({ data: [{ ...tier }] })
       } catch (e) {
         const message = e.message || e.cause.message
         this.showNotification({ message, color: 'red' })
@@ -191,10 +196,10 @@ export default {
   computed: {
     ...mapGetters('dao', ['selectedDao']),
 
-    hasLevels () { return this?.levels && this?.levels?.length > 0 },
     hasRoles () { return this?.roles && this?.roles?.length > 0 },
-    levelCount () { return this?.levels ? this?.levels?.length : 0 },
-    roleCount () { return this.roles ? this.roles?.length : 0 }
+    hasTiers () { return this?.tiers && this?.tiers?.length > 0 },
+    roleCount () { return this.roles ? this.roles?.length : 0 },
+    tierCount () { return this?.tiers ? this?.tiers?.length : 0 }
   },
 
   watch: {
@@ -220,12 +225,14 @@ export default {
       v-model="tab"
     )
       q-tab(:name="TABS.TYPE" :label="$t('configuration.settings-structure.roles.tabs.types')" :ripple="false")
-      q-tab(:name="TABS.LEVEL" :label="$t('configuration.settings-structure.roles.tabs.levels')" :ripple="false")
+      q-tab(:name="TABS.TIER" :label="$t('configuration.settings-structure.roles.tabs.tiers')" :ripple="false")
 
     section(v-if="tab === TABS.TYPE").q-mt-md
       header.column.justify-center.text-center.items-center(v-if="!hasRoles && state === STATES.WAITING")
+        q-icon.q-mt-xl(name="fas fa-briefcase" size="60px")
         p.text-sm.text-h-gray.leading-loose.q-mt-md {{ $t('configuration.settings-structure.roles.type.heading') }}
         q-btn.q-px-xl.text-bold(
+          :disable="!isAdmin"
           :label="$t('configuration.settings-structure.roles.type.nav.create')"
           @click="state = STATES.CREATING_ROLE"
           color="primary"
@@ -233,7 +240,6 @@ export default {
           no-caps
           rounded
           unelevated
-          v-if="isAdmin"
         )
 
       section(v-if="hasRoles").row.q-col-gutter-md
@@ -262,7 +268,7 @@ export default {
                 p.q-pa-none.q-ma-none.text-sm.text-primary.text-bold.leading-loose {{ role?.assignmentAggregate?.count || 0 }} {{ $t('dao.member') }}
 
       section(v-if="state === STATES.CREATING_ROLE")
-        .hr.q-my-xl(v-if="hasLevels")
+        .hr.q-my-xl(v-if="hasRoles")
 
         .full-width
             label.h-label {{ $t('configuration.settings-structure.roles.type.form.name.label') }}
@@ -330,24 +336,25 @@ export default {
           v-if="isAdmin"
         )
 
-    section(v-if="tab === TABS.LEVEL").q-mt-md
-      header.column.justify-center.text-center.items-center(v-if="!hasLevels && state !== STATES.WAITING")
-        p.text-sm.text-h-gray.leading-loose.q-mt-md {{ $t('configuration.settings-structure.roles.level.heading') }}
+    section(v-if="tab === TABS.TIER").q-mt-md
+      header.column.justify-center.text-center.items-center(v-if="!hasTiers && state === STATES.WAITING")
+        q-icon.q-mt-xl(name="fas fa-chart-bar" size="60px")
+        p.text-sm.text-h-gray.leading-loose.q-mt-md {{ $t('configuration.settings-structure.roles.tier.heading') }}
         q-btn.q-px-xl.text-bold(
-          :label="$t('configuration.settings-structure.roles.level.nav.create')"
-          @click="state = STATES.CREATING_LEVEL"
+          :disable="!isAdmin"
+          :label="$t('configuration.settings-structure.roles.tier.nav.create')"
+          @click="state = STATES.CREATING_TIER"
           color="primary"
           icon="fas fa-plus"
           no-caps
           rounded
           unelevated
-          v-if="isAdmin"
         )
 
-      section(v-if="hasLevels").row.q-col-gutter-md
-        template(v-for="level in levels")
-          article.col-12(:class="['col-md-'+ Math.min(Math.max((12/levelCount), 3), 12)]")
-            widget(:title="level.name" shadow bar)
+      section(v-if="hasTiers").row.q-col-gutter-md
+        template(v-for="tier in tiers")
+          article.col-12(:class="['col-md-'+ Math.min(Math.max((12/tierCount), 3), 12)]")
+            widget(:title="tier.name" shadow bar)
               template(v-slot:header)
                 q-btn.q-pa-xs.relative-position(
                   color="primary"
@@ -360,25 +367,25 @@ export default {
                 )
                   q-menu
                     q-list(dense)
-                      q-item(@click="_deleteLevel(level)" clickable v-close-popup)
+                      q-item(@click="_deleteLevel(tier)" clickable v-close-popup)
                         q-item-section {{ $t('actions.delete') }}
 
-              p.q-pa-none.q-ma-none.text-sm.text-h-gray.leading-loose.q-mt-xs ${{ formatCurrency(level.annualAmount) }}/year
+              p.q-pa-none.q-ma-none.text-sm.text-h-gray.leading-loose.q-mt-xs ${{ formatCurrency(tier.annualAmount) }}/year
               .hr.q-my-xl
               .row.items-center
                 q-avatar.q-mr-sm.bg-h-gray(size="md" text-color="white" icon="fas fa-user")
-                p.q-pa-none.q-ma-none.text-sm.text-primary.text-bold.leading-loose {{ level?.assignmentAggregate?.count || 0 }} {{ $t('dao.member') }}
+                p.q-pa-none.q-ma-none.text-sm.text-primary.text-bold.leading-loose {{ tier?.assignmentAggregate?.count || 0 }} {{ $t('dao.member') }}
 
-      section(v-if="state === STATES.CREATING_LEVEL")
-        .hr.q-my-xl(v-if="hasLevels")
+      section(v-if="state === STATES.CREATING_TIER")
+        .hr.q-my-xl(v-if="hasTiers")
 
         .row.q-col-gutter-md
           .col-3
-            label.h-label {{ $t('configuration.settings-structure.roles.level.form.name.label') }}
+            label.h-label {{ $t('configuration.settings-structure.roles.tier.form.name.label') }}
             q-input.q-my-xs(
               :debounce="200"
               :disable="!isAdmin"
-              :placeholder="$t('configuration.settings-structure.roles.level.form.name.placeholder')"
+              :placeholder="$t('configuration.settings-structure.roles.tier.form.name.placeholder')"
               bg-color="white"
               color="accent"
               dense
@@ -386,15 +393,15 @@ export default {
               outlined
               ref="name"
               rounded
-              v-model='level.name'
+              v-model='tier.name'
             )
 
           .col-3
-            label.h-label {{ $t('configuration.settings-structure.roles.level.form.yearly-reward.label') }}
+            label.h-label {{ $t('configuration.settings-structure.roles.tier.form.yearly-reward.label') }}
             q-input.q-my-xs(
               :debounce="200"
               :disable="!isAdmin"
-              :placeholder="$t('configuration.settings-structure.roles.level.form.yearly-reward.placeholder')"
+              :placeholder="$t('configuration.settings-structure.roles.tier.form.yearly-reward.placeholder')"
               bg-color="white"
               color="accent"
               dense
@@ -403,17 +410,17 @@ export default {
               ref="annualAmount"
               rounded
               type='number'
-              v-model='level.annualAmount'
+              v-model='tier.annualAmount'
             )
               template(v-slot:before)
                 q-icon(name="fas fa-dollar-sign")
 
           .col-3
-            label.h-label {{ $t('configuration.settings-structure.roles.level.form.montly-reward.label') }}
+            label.h-label {{ $t('configuration.settings-structure.roles.tier.form.montly-reward.label') }}
             q-input.q-my-xs(
               :debounce="200"
               :disable="true"
-              :value='parseFloat(level.annualAmount / 12).toFixed(2)'
+              :value='parseFloat(tier.annualAmount / 12).toFixed(2)'
               bg-color="white"
               color="accent"
               dense
@@ -427,7 +434,7 @@ export default {
                 q-icon(name="fas fa-dollar-sign")
 
           .col-3
-            label.h-label {{ $t('configuration.settings-structure.roles.level.form.min-deferred.label') }}
+            label.h-label {{ $t('configuration.settings-structure.roles.tier.form.min-deferred.label') }}
             .row.full-width.items-center
               .col.row.q-mr-sm
                 q-slider(
@@ -435,7 +442,7 @@ export default {
                   :min="0"
                   :step="1"
                   color="primary"
-                  v-model="level.minDeferred"
+                  v-model="tier.minDeferred"
                 )
               .col-3
                 q-input.rounded-border.q-py-sm(
@@ -444,12 +451,12 @@ export default {
                   outlined
                   rounded
                   suffix="%"
-                  v-model.number="level.minDeferred"
+                  v-model.number="tier.minDeferred"
                 )
 
         nav.full-width.q-my-xl.row.justify-end
             q-btn.col-auto.q-px-xl.rounded-border.text-bold.q-mr-xs(
-              :label="$t('configuration.settings-structure.roles.level.form.cancel')"
+              :label="$t('configuration.settings-structure.roles.tier.form.cancel')"
               @click="state = STATES.WAITING"
               color="white"
               no-caps
@@ -458,7 +465,7 @@ export default {
               unelevated
             )
             q-btn.col-auto.q-px-xl.rounded-border.text-bold.q-ml-xs(
-              :label="$t('configuration.settings-structure.roles.level.form.submit')"
+              :label="$t('configuration.settings-structure.roles.tier.form.submit')"
               @click="_createLevel"
               color="secondary"
               no-caps
@@ -466,10 +473,10 @@ export default {
               unelevated
             )
 
-      nav.full-width.row.justify-end.q-mt-xl(v-if="hasLevels && state === STATES.WAITING")
+      nav.full-width.row.justify-end.q-mt-xl(v-if="hasTiers && state === STATES.WAITING")
         q-btn.q-px-xl.text-bold(
-          :label="$t('configuration.settings-structure.roles.level.nav.create')"
-          @click="state = STATES.CREATING_LEVEL"
+          :label="$t('configuration.settings-structure.roles.tier.nav.create')"
+          @click="state = STATES.CREATING_TIER"
           color="primary"
           icon="fas fa-plus"
           no-caps